requirements.txt
streamlit==1.27.2
langchain==0.2.3
langchain-openai==0.1.9
langchain_community==0.2.4
openai==1.39.0
pinecone-client==5.0.1
pinecone-text==0.9.0
pillow==9.5.0
tqdm==4.66.1
cohere==5.5.6
unstructured==0.10.18